Use float for speed values from curves and fan controllers
They're between 0.0 and 255.0 and then get normalized to a value
between 0 (or MinPWM, if shouldNeverStop) and MaxPWM.
Only that normalized value is an actual (integer) PWM value, so there's
no reason to round to float in any step before that normalization,
that only decreases accuracy.

Because of this I also changed "PWM" to "Speed" in a few places

+++ internal/curves/curve.go
@@ -13,9 +13,9 @@ type SpeedCurve interface {
 	GetId() string
 	// Evaluate update the value of this SpeedCurve, by calculates a new value based on the current sensor values
 	// returns a value in [0..255]
-	Evaluate() (value int, err error)
+	Evaluate() (value float64, err error)
 	// CurrentValue returns the current value of the curve, which was calculated by the previous call to Evaluate
-	CurrentValue() int
+	CurrentValue() float64
 }
 
 var (
